refactor: create CourseListCode enum
