+++ src/components/sections/dictionary/courselist/CourseListSection.jsx
@@ -6,7 +6,7 @@ import ReactGA from 'react-ga4';
 
 import { appBoundClassNames as classNames } from '../../../../common/boundClassNames';
 
-import { CourseListCode } from '../../../../reducers/dictionary/list';
+import { CourseListCode } from '@/shapes/enum';
 
 import Scroller from '../../../Scroller';
 import CourseSearchSubSection from './CourseSearchSubSection';
