+++ src/components/sections/dictionary/courselist/CourseListTabs.jsx
@@ -7,7 +7,7 @@ import ReactGA from 'react-ga4';
 
 import { appBoundClassNames as classNames } from '../../../../common/boundClassNames';
 
-import { CourseListCode } from '../../../../reducers/dictionary/list';
+import { CourseListCode } from '@/shapes/enum';
 
 import { openSearch, closeSearch } from '../../../../actions/dictionary/search';
 import { setSelectedListCode, setListCourses } from '../../../../actions/dictionary/list';
