+++ src/components/sections/planner/courselist/CourseListSection.jsx
@@ -6,7 +6,7 @@ import ReactGA from 'react-ga4';
 
 import { appBoundClassNames as classNames } from '../../../../common/boundClassNames';
 
-import { CourseListCode } from '../../../../reducers/planner/list';
+import { CourseListCode } from '@/shapes/enum';
 
 import Scroller from '../../../Scroller';
 import CourseSearchSubSection from './CourseSearchSubSection';
