+++ src/components/sections/planner/courselist/CourseSearchSubSection.jsx
@@ -6,7 +6,7 @@ import axios from 'axios';
 import { debounce } from 'lodash';
 
 import { appBoundClassNames as classNames } from '../../../../common/boundClassNames';
-import { CourseListCode } from '../../../../reducers/planner/list';
+import { CourseListCode } from '@/shapes/enum';
 
 import Divider from '../../../Divider';
 import SearchFilter from '../../../SearchFilter';
